Add assignment attachment upload/download support
Introduce a new assignment_attachment_controller to handle saving, listing, and downloading attachments. Made 5 small tests in  separate test file for attachment, allows for multiple file uploads.
