+++ flask_backend/tests/test_assignments.py
@@ -8,8 +8,6 @@ Tests for assignments endpoints
 
 import json
 import datetime
-import io
-import os
 
 def test_teacher_can_create_assignment(test_client, make_admin):
     """
@@ -83,194 +81,6 @@ def test_create_assignment_missing_fields(test_client, make_admin):
     assert response.status_code == 400
     assert response.json["msg"] == "Assignment name is required"
 
-
-def test_teacher_can_create_assignment_with_multipart_without_attachments(test_client, make_admin):
-    """Ensure multipart form submissions still create assignments successfully."""
-    make_admin(email="admin@example.com", password="admin", name="adminuser")
-
-    test_client.post(
-        "/auth/login",
-        data=json.dumps({"email": "admin@example.com", "password": "admin"}),
-        headers={"Content-Type": "application/json"},
-    )
-
-    class_response = test_client.post(
-        "/class/create_class",
-        data=json.dumps({"name": "History 201"}),
-        headers={"Content-Type": "application/json"},
-    )
-    class_id = class_response.json["class"]["id"]
-
-    response = test_client.post(
-        "/assignment/create_assignment",
-        data={
-            "courseID": str(class_id),
-            "name": "Essay 2",
-            "rubric": "Clarity",
-            "due_date": datetime.datetime(2027, 1, 10, 12, 0, 0).isoformat(),
-            "start_date": datetime.datetime(2027, 1, 1, 12, 0, 0).isoformat(),
-        },
-        content_type="multipart/form-data",
-    )
-
-    assert response.status_code == 201
-    assert response.json["assignment"]["name"] == "Essay 2"
-    assert response.json["attachments"] == []
-
-
-def test_teacher_can_create_assignment_with_attachment(test_client, make_admin):
-    """Ensure assignment creation supports optional file attachments."""
-    make_admin(email="teacher@example.com", password="teacher", name="teacheruser")
-
-    test_client.post(
-        "/auth/login",
-        data=json.dumps({"email": "teacher@example.com", "password": "teacher"}),
-        headers={"Content-Type": "application/json"},
-    )
-
-    class_response = test_client.post(
-        "/class/create_class",
-        data=json.dumps({"name": "CompSci 101"}),
-        headers={"Content-Type": "application/json"},
-    )
-    class_id = class_response.json["class"]["id"]
-
-    response = test_client.post(
-        "/assignment/create_assignment",
-        data={
-            "courseID": str(class_id),
-            "name": "Project Spec",
-            "rubric": "Completeness",
-            "attachments": (io.BytesIO(b"file-contents"), "spec.txt"),
-        },
-        content_type="multipart/form-data",
-    )
-
-    assert response.status_code == 201
-    assert len(response.json["attachments"]) == 1
-    assert response.json["attachments"][0]["original_name"] == "spec.txt"
-
-    assignment_id = response.json["assignment"]["id"]
-    stored_name = response.json["attachments"][0]["stored_name"]
-    expected_path = os.path.join(
-        test_client.application.instance_path,
-        "assignment_uploads",
-        str(assignment_id),
-        stored_name,
-    )
-    assert os.path.exists(expected_path)
-
-
-def test_teacher_can_create_assignment_with_attachment_file_field_name(test_client, make_admin):
-    """Support clients that submit a single file under field name 'file'."""
-    make_admin(email="teacher2@example.com", password="teacher", name="teacheruser2")
-
-    test_client.post(
-        "/auth/login",
-        data=json.dumps({"email": "teacher2@example.com", "password": "teacher"}),
-        headers={"Content-Type": "application/json"},
-    )
-
-    class_response = test_client.post(
-        "/class/create_class",
-        data=json.dumps({"name": "CompSci 102"}),
-        headers={"Content-Type": "application/json"},
-    )
-    class_id = class_response.json["class"]["id"]
-
-    response = test_client.post(
-        "/assignment/create_assignment",
-        data={
-            "courseID": str(class_id),
-            "name": "Project Plan",
-            "file": (io.BytesIO(b"plan-contents"), "plan.txt"),
-        },
-        content_type="multipart/form-data",
-    )
-
-    assert response.status_code == 201
-    assert len(response.json["attachments"]) == 1
-    assert response.json["attachments"][0]["original_name"] == "plan.txt"
-
-
-def test_get_assignments_includes_attachment_metadata(test_client, make_admin):
-    make_admin(email="teacher3@example.com", password="teacher", name="teacheruser3")
-
-    test_client.post(
-        "/auth/login",
-        data=json.dumps({"email": "teacher3@example.com", "password": "teacher"}),
-        headers={"Content-Type": "application/json"},
-    )
-
-    class_response = test_client.post(
-        "/class/create_class",
-        data=json.dumps({"name": "CompSci 103"}),
-        headers={"Content-Type": "application/json"},
-    )
-    class_id = class_response.json["class"]["id"]
-
-    create_response = test_client.post(
-        "/assignment/create_assignment",
-        data={
-            "courseID": str(class_id),
-            "name": "Project Summary",
-            "attachments": (io.BytesIO(b"summary"), "summary.txt"),
-        },
-        content_type="multipart/form-data",
-    )
-
-    assignment_id = create_response.json["assignment"]["id"]
-    list_response = test_client.get(f"/assignment/{class_id}")
-
-    assert list_response.status_code == 200
-    match = next((a for a in list_response.json if a["id"] == assignment_id), None)
-    assert match is not None
-    assert match["attachments"]
-    summary_attachment = next(
-        (a for a in match["attachments"] if a["original_name"] == "summary.txt"),
-        None,
-    )
-    assert summary_attachment is not None
-    assert summary_attachment["download_url"].startswith(
-        f"/assignment/{assignment_id}/attachment/"
-    )
-
-
-def test_teacher_can_download_assignment_attachment(test_client, make_admin):
-    make_admin(email="teacher4@example.com", password="teacher", name="teacheruser4")
-
-    test_client.post(
-        "/auth/login",
-        data=json.dumps({"email": "teacher4@example.com", "password": "teacher"}),
-        headers={"Content-Type": "application/json"},
-    )
-
-    class_response = test_client.post(
-        "/class/create_class",
-        data=json.dumps({"name": "CompSci 104"}),
-        headers={"Content-Type": "application/json"},
-    )
-    class_id = class_response.json["class"]["id"]
-
-    create_response = test_client.post(
-        "/assignment/create_assignment",
-        data={
-            "courseID": str(class_id),
-            "name": "Download Test",
-            "attachments": (io.BytesIO(b"download-body"), "download.txt"),
-        },
-        content_type="multipart/form-data",
-    )
-
-    assignment_id = create_response.json["assignment"]["id"]
-    stored_name = create_response.json["attachments"][0]["stored_name"]
-
-    download_response = test_client.get(
-        f"/assignment/{assignment_id}/attachment/{stored_name}"
-    )
-
-    assert download_response.status_code == 200
-    assert download_response.data == b"download-body"
 
 def test_non_assigned_teacher_cannot_create_assignment(test_client, make_admin):
     """
